api/requirements.txt
flask
python-dotenv
gunicorn
pre-commit
wheel
flask-sqlalchemy
psycopg2-binary
flask-migrate
flask-script
mysqlclient
